fix(core): align movement calls and harden numeric conversions across server code
- add explicit static_cast conversions for timers, cooldowns, counters and comparisons
- update MoveJump usages to the current function signature
- fix unsafe/incorrect printf format specifiers for 64-bit values
- normalize project file headers to DestinyCore
- resolve minor logic issues in power/health handling and boss script calculations

+++ src/server/game/Battlefield/Battlefield.cpp
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2008-2010 TrinityCore <http://www.trinitycore.org/>
- * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -1097,7 +1096,7 @@ bool BfCapturePoint::Update(uint32 diff)
                 HandlePlayerEnter(*itr);
 
     // get the difference of numbers
-    float fact_diff = ((float)m_activePlayers[0].size() - (float)m_activePlayers[1].size()) * diff / BATTLEFIELD_OBJECTIVE_UPDATE_INTERVAL;
+    float fact_diff = ((float)m_activePlayers[0].size() - (float)m_activePlayers[1].size()) * diff / static_cast<float>(BATTLEFIELD_OBJECTIVE_UPDATE_INTERVAL);
     if (!fact_diff)
         return false;
