+++ src/server/game/Instances/InstanceScript.cpp
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2008-2012 TrinityCore <http://www.trinitycore.org/>
- * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -1155,8 +1154,8 @@ void InstanceScript::StartCombatResurrection()
     auto timer = uint32(value / 100.0f);
 
     value -= static_cast<float>(timer) * 100.0f;
-    timer *= MINUTE * IN_MILLISECONDS;
-    value *= MINUTE / 100.0f * IN_MILLISECONDS;
+    timer *= static_cast<uint32>(MINUTE) * static_cast<uint32>(IN_MILLISECONDS);
+    value *= static_cast<float>(MINUTE) / 100.0f * static_cast<float>(IN_MILLISECONDS);
     timer += uint32(value);
 
     _maxInCombatResCount = instance->IsRaid() ? 9 : 0;
