+++ src/server/game/Spells/Auras/SpellAuraEffects.cpp
@@ -9368,9 +9368,9 @@ void AuraEffect::HandleModTimeRate(AuraApplication const* aurApp, uint8 mode, bo
             continue;
 
         double t = getPreciseTime();
-        double cooldown = double(itr.second.end > t ? itr.second.end - t : 0.0) * IN_MILLISECONDS;
+        double cooldown = double(itr.second.end > t ? itr.second.end - t : 0.0) * static_cast<double>(IN_MILLISECONDS);
         cooldown *= Multiplier;
-        itr.second.end = getPreciseTime() + cooldown / IN_MILLISECONDS;
+        itr.second.end = getPreciseTime() + cooldown / static_cast<double>(IN_MILLISECONDS);
 
         WorldPackets::Spells::ModifyCooldownRecoverySpeed package;
         package.SpellID = spellInfo->Id;
@@ -9559,9 +9559,9 @@ void AuraEffect::HandleAuraModCooldownSpeedRate(AuraApplication const* auraApp, 
         ApplyPercentModFloatVar(Multiplier, amount, !apply);
 
         double t = getPreciseTime();
-        double cooldown = double(itr.second.end > t ? itr.second.end - t : 0.0) * IN_MILLISECONDS;
+        double cooldown = double(itr.second.end > t ? itr.second.end - t : 0.0) * static_cast<double>(IN_MILLISECONDS);
         cooldown *= Multiplier;
-        itr.second.end = getPreciseTime() + cooldown / IN_MILLISECONDS;
+        itr.second.end = getPreciseTime() + cooldown / static_cast<double>(IN_MILLISECONDS);
 
         WorldPackets::Spells::ModifyCooldownRecoverySpeed package;
         package.SpellID = spellInfo->Id;
@@ -9663,10 +9663,10 @@ void AuraEffect::HandleAuraModCooldownPct(AuraApplication const* auraApp, uint8 
             }
 
             double t = getPreciseTime();
-            double cooldown = double(itr.second.end > t ? itr.second.end - t : 0.0) * IN_MILLISECONDS;
+            double cooldown = double(itr.second.end > t ? itr.second.end - t : 0.0) * static_cast<double>(IN_MILLISECONDS);
             cooldown = CalculatePct(cooldown, (100 - (itr.second.rate * 100)) + 100); // Normalize CD
             cooldown *= modCooldownPct;
-            itr.second.end = getPreciseTime() + cooldown / IN_MILLISECONDS;
+            itr.second.end = getPreciseTime() + cooldown / static_cast<double>(IN_MILLISECONDS);
             itr.second.rate = modCooldownPct;
 
             WorldPackets::Character::CooldownCheat package;
